Adding Lagermeister code
Initial commit of the Lagermeister arduino code.  Currently it works but has some dead code, some non-optimal functions, and it's disorganized.  I might organize it and improve it (especially the function to determine whether to turn on the heater or cooler when it's close to the right temperature), but as it works in its current form, my motivation is somewhat lacking at the moment.